[red-knot] split call-outcome enums to their own submodule
